# src/modules/payments/dto/request/update-payment.dto.ts
import { ApiPropertyOptional } from "@nestjs/swagger";

export class UpdatePaymentDto {
	@ApiPropertyOptional({ minimum: 1 })
	amount?: number;

	@ApiPropertyOptional({ type: "object" })
	data?: Record<string, unknown>;
}